correct ts error and move files to correct directory
