+++ examples/components/Block.vue
@@ -1,0 +1,14 @@
+<template>
+    <h1>HELLO</h1>
+</template>
+
+<script>
+export default {
+    name: 'Block'
+}
+</script>
+
+<docs>
+    <Block />
+</docs>
+
